LayoutTests/fast/forms/novalidate-attribute-expected.txt

This test performs some simple check on the novalidate attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS f.hasAttribute('novalidate') is false
PASS f.getAttribute('novalidate') is null
PASS f.novalidate is false
PASS f.hasAttribute('novalidate') is true
PASS f.getAttribute('novalidate') is ''
PASS f.novalidate is true
PASS f.hasAttribute('novalidate') is false
PASS f.getAttribute('novalidate') is null
PASS f.novalidate is false
PASS f2.hasAttribute('novalidate') is true
PASS f2.getAttribute('novalidate') is ''
PASS f2.novalidate is true
PASS f2.novalidate is false
PASS f2.hasAttribute('novalidate') is true
PASS f2.getAttribute('novalidate') is ''
PASS f2.novalidate is true
PASS successfullyParsed is true

TEST COMPLETE
